refactor: tuist 공용 컬러셋  이용하게 수정
